+++ 20150926/tono_serarin/characters/monsters/chimera.rb
@@ -1,0 +1,9 @@
+class Chimera < Monster
+  def initialize
+    self.name = "キメラ"
+    self.attack_power = 7
+    self.defense_power = 3
+    self.hit_point = 6
+    self.experience_point = 5
+  end
+end
