+++ 20150926/tono_serarin/characters/monsters/dragon.rb
@@ -1,0 +1,9 @@
+class Dragon < Monster
+  def initialize
+    self.name = "ドラゴン"
+    self.attack_power = 15
+    self.defense_power = 5
+    self.hit_point = 20
+    self.experience_point = 100
+  end
+end
