+++ 20150926/tono_serarin/characters/monsters/slime.rb
@@ -1,0 +1,9 @@
+class Slime < Monster
+  def initialize
+    self.name = "スライム"
+    self.attack_power = 5
+    self.defense_power = 3
+    self.hit_point = 3
+    self.experience_point = 1
+  end
+end
