+++ 20150926/tono_serarin/draque.rb
@@ -1,3 +1,9 @@
+require "./characters/hero"
+require "./characters/monster"
+require "./characters/monsters/slime"
+require "./characters/monsters/chimera"
+require "./characters/monsters/dragon"
+
 class Game
 
   def initialize
@@ -40,7 +46,7 @@ class Game
         if @monster.hit_point <= 0
           puts "#{@monster.name}は倒れた。"
           @hero.experience_point += @monster.experience_point
-          puts "経験値を#{@hero.experience_point}手に入れた！"          
+          puts "経験値を#{@hero.experience_point}手に入れた！"
           break
         end
 
@@ -59,101 +65,5 @@ class Game
     end
   end
 end
-
-class Hero
-  attr_accessor :level, :attack_power, :defense_power, :hit_point, :magic_power, :experience_point
-
-  def initialize
-    self.level = 1
-    self.attack_power = 4
-    self.defense_power = 4
-    self.hit_point = 15
-    self.magic_power = 0
-    self.experience_point = 0
-  end
-
-  def attack(monster)
-  	if (1..32).to_a.sample == 1
-  		critical_hit
-  	else
-  		normal_attack(monster)
-  	end
-  end
-
-  def normal_attack(monster)
-    a = (attack_power - monster.defense_power) / 2
-    if a < 2
-      [1,2].sample
-    else
-      ((a+1) * (0..255).to_a.sample / 256 + a) / 2
-    end
-  end
-
-  def critical_hit
-    puts "会心の一撃！！！！"
-		attack_power - (attack_power / 2) * (0..255).to_a.sample / 256
-	end
-end
-
-class Monster
-	attr_accessor :name, :attack_power, :defense_power, :hit_point, :experience_point
-
-  def initialize
-		self.name = "スライム"
-		self.attack_power = 5
-		self.defense_power = 3
-		self.hit_point = 3
-		self.experience_point = 1
-  end
-
-  def attack(hero)
-    # Ａ＝（敵の攻撃力＊２－勇者の守備力）／２
-    a = (attack_power * 2 - hero.defense_power) / 2
-    b = (attack_power / 2) + 1
-    if a <= 0
-      [1,2].sample
-    else
-      random = (0..255).to_a.sample
-      if a < b
-        #ダメージは、｛２＋（敵の攻撃力／２＋１）＊（０～２５５）／２５６｝／３
-        (b * random  / 256 + 2) / 3
-      else
-        # Ａ＞敵の攻撃力／２＋１の場合、ダメージは、｛Ａ＋（Ａ＋１）＊（０～２５５）／２５６｝／２
-        ((a + 1) * random / 256 + 2) / 2
-      end
-    end
-  end
-end
-
-class Slime < Monster
-  def initialize
-    self.name = "スライム"
-    self.attack_power = 5
-    self.defense_power = 3
-    self.hit_point = 3
-    self.experience_point = 1
-  end
-end
-
-class Chimera < Monster
-  def initialize
-    self.name = "キメラ"
-    self.attack_power = 7
-    self.defense_power = 3
-    self.hit_point = 6
-    self.experience_point = 5
-  end
-end
-
-class Dragon < Monster
-  def initialize
-    self.name = "ドラゴン"
-    self.attack_power = 15
-    self.defense_power = 5
-    self.hit_point = 20
-    self.experience_point = 100
-  end
-end
-
 
 Game.new
